consolidated newsletter tracking into pages/_app.js
