+++ pages/index.js
@@ -1,5 +1,4 @@
-import React, { useEffect } from 'react';
-import { useAnalytics } from '../lib/hooks/useAnalytics.js';
+import React from 'react';
 import { hasuraStreamArticles } from '../lib/homepage.js';
 import { cachedContents } from '../lib/cached';
 import { hasuraGetHomepageEditor } from '../lib/articles.js';
@@ -9,23 +8,6 @@ import Homepage from '../components/Homepage';
 import LandingPage from '../components/LandingPage';
 
 export default function Home(props) {
-  const {
-    setDimension,
-    trackPageViewedWithDimension,
-    trackMailChimpParams,
-  } = useAnalytics();
-  useEffect(() => {
-    let isSubscriber = trackMailChimpParams();
-    if (isSubscriber) {
-      setDimension('dimension5', true);
-      trackPageViewedWithDimension(
-        window.location.pathname + window.location.search,
-        'dimension5',
-        true
-      );
-    }
-  }, []);
-
   const component = props.siteMetadata.landingPage ? (
     <LandingPage {...props} />
   ) : (
